% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{read_mod}
\alias{read_mod}
\title{Read in NONMEM model file for pharmpy}
\usage{
read_mod(dir, modfile)
}
\arguments{
\item{dir}{Character string with the path to the NONMEM model file}

\item{modfile}{Character string with name of the model file}
}
\value{
Returns the pharmr model object
}
\description{
Read in NONMEM model file for pharmpy
}
\examples{
\dontrun{
read_mod(dir = './01_NONMEM/01_BaseModel/', modfile = "run026.mod")
}
}
