% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{run_nonmem}
\alias{run_nonmem}
\title{Fit defined NONMMEM model according to the current pharmpy functions.}
\usage{
run_nonmem(model, dir)
}
\arguments{
\item{model}{pharmpy model object}

\item{dir}{Character string providing the output directory (folder) where
the modelfit content will be saved.}
}
\value{
Returns the results from pharmpr::fit()
}
\description{
Fit defined NONMMEM model according to the current pharmpy functions.
}
\examples{
\dontrun{
run_nonmem(model = model)
}
}
